+++ cryptofeed/bitstamp/bitstamp.py
@@ -6,6 +6,7 @@ associated with this software.
 '''
 import json
 import asyncio
+import logging
 from decimal import Decimal
 
 import requests
@@ -14,8 +15,10 @@ from sortedcontainers import SortedDict as sd
 from cryptofeed.exchanges import BITSTAMP
 from cryptofeed.feed import Feed
 from cryptofeed.defines import BID, ASK, TRADES, L3_BOOK
-from cryptofeed.callback import Callback
 from cryptofeed.standards import pair_exchange_to_std, pair_std_to_exchange
+
+
+LOG = logging.getLogger('feedhandler')
 
 
 class Bitstamp(Feed):
@@ -36,8 +39,7 @@ class Bitstamp(Feed):
         loop = asyncio.get_event_loop()
         btc_usd_url = 'https://www.bitstamp.net/api/order_book/'
         url = 'https://www.bitstamp.net/api/v2/order_book/{}/'
-        pairs =  [pair_std_to_exchange(pair, self.id) for pair in self.pairs]
-        futures = [loop.run_in_executor(None, requests.get, url.format(pair) if pair != 'BTC-USD' else btc_usd_url) for pair in pairs]
+        futures = [loop.run_in_executor(None, requests.get, url.format(pair) if pair != 'BTC-USD' else btc_usd_url) for pair in self.pairs]
 
         results = []
         for future in futures:
@@ -46,6 +48,7 @@ class Bitstamp(Feed):
 
         for res, pair in zip(results, self.pairs):
             orders = res.json()
+            pair = pair_exchange_to_std(pair)
             self.book[pair] = {BID: sd(), ASK: sd()}
             self.seq_no[pair] = orders['timestamp']
 
@@ -113,20 +116,20 @@ class Bitstamp(Feed):
         msg = msg.replace("\\", '')
         msg = msg.replace("\"{", "{")
         msg = msg.replace("}\"", "}")
-        msg = json.loads(msg)
+        msg = json.loads(msg, parse_float=Decimal)
         if 'pusher' in msg['event']:
             if msg['event'] == 'pusher:connection_established':
                 pass
             elif msg['event'] == 'pusher_internal:subscription_succeeded':
                 pass
             else:
-                print("Unexpected pusher message {}".format(msg))
+                LOG.warning("{} - Unexpected pusher message {}".format(self.id, msg))
         elif msg['event'] == 'trade':
             await self._trades(msg)
         elif msg['event'] == 'data':
             await self._order_book(msg)
         else:
-            print('Invalid message type {}'.format(msg))
+            LOG.warning("{} - Invalid message type {}".format(self.id, msg))
 
     async def subscribe(self, websocket):
         # if channel is order book we need to subscribe to the diff channel
@@ -135,7 +138,6 @@ class Bitstamp(Feed):
         # are pre-timestamp on the response from the REST endpoint
         for channel in self.channels:
             for pair in self.pairs:
-                pair = pair_std_to_exchange(pair, self.id)
                 await websocket.send(
                     json.dumps({
                         "event": "pusher:subscribe",