+++ cryptofeed/poloniex/poloniex.py
@@ -5,6 +5,7 @@ Please see the LICENSE file for the terms and conditions
 associated with this software.
 '''
 import json
+import logging
 from decimal import Decimal
 
 from sortedcontainers import SortedDict as sd
@@ -17,12 +18,16 @@ from cryptofeed.exchanges import POLONIEX
 from .pairs import poloniex_id_pair_mapping
 
 
+LOG = logging.getLogger('feedhandler')
+
+
 class Poloniex(Feed):
     id = POLONIEX
 
     def __init__(self, pairs=None, channels=None, callbacks=None):
         if pairs:
-            raise ValueError("Poloniex does not support pairs on a channel")
+            LOG.error("Poloniex does not support pairs")
+            raise ValueError("Poloniex does not support pairs")
 
         super(Poloniex, self).__init__('wss://api2.poloniex.com',
                                      channels=channels,
@@ -32,7 +37,7 @@ class Poloniex(Feed):
         # currencyPair, last, lowestAsk, highestBid, percentChange, baseVolume,
         # quoteVolume, isFrozen, 24hrHigh, 24hrLow
         pair_id, _, ask, bid, _, _, _, _, _, _ = msg
-        pair = poloniex_id_pair_mapping[pair_id]
+        pair = pair_exchange_to_std(poloniex_id_pair_mapping[pair_id])
         await self.callbacks[TICKER](feed=self.id,
                                      pair=pair,
                                      bid=Decimal(bid),
@@ -90,13 +95,14 @@ class Poloniex(Feed):
                                                  amount=amount,
                                                  price=price)
                 else:
-                    print("Unexpected message received: {}".format(msg))
+                    LOG.warning("{} - Unexpected message received: {}".format(self.id, msg))
+
         await self.callbacks[L3_BOOK](feed=self.id, pair=pair, book=self.l3_book[pair])
 
     async def message_handler(self, msg):
-        msg = json.loads(msg)
+        msg = json.loads(msg, parse_float=Decimal)
         if 'error' in msg:
-            print("Error from exchange: {}".format(msg))
+            LOG.error("{} - Error from exchange: {}".format(self.id, msg))
             return
 
         chan_id = msg[0]
@@ -121,13 +127,10 @@ class Poloniex(Feed):
             # heartbeat - ignore
             pass
         else:
-            print('Invalid message type {}'.format(msg))
+            LOG.warning('{} - Invalid message type {}'.format(self.id, msg))
 
     async def subscribe(self, websocket):
         for channel in self.channels:
-            chan = pair_std_to_exchange(channel, POLONIEX)
-            if chan:
-                channel = chan
             await websocket.send(json.dumps({"command": "subscribe",
                                              "channel": channel
                                             }))